requirements.txt
langchain>=0.0.300
langchain-community>=0.0.10
jira>=3.5.1
python-dotenv>=1.0.0
chromadb>=0.4.15
tiktoken>=0.5.1
pandas>=2.1.0
numpy>=1.24.0
python-jose>=3.3.0
beautifulsoup4>=4.12.2
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
openai>=1.3.0
sentence-transformers>=2.2.2
